+++ lib/src/extension/common/index.dart
@@ -1,2 +1,3 @@
 export 'list.dart';
+export 'path.dart';
 export 'string.dart';