+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.0.3:
+feat:
+1. rootDir 
+2. Breadpoint: some option cloud be null, set with a default as protect.
+3. fromCharCodes, switch `List<int>` to String
+4. add test usecase
+5. ignore test rootDir
+
 # 0.0.2
 test: :white_check_mark: test option is derive, it could switch
 doc: Feature about args